add test_gcs_fault_tolerance.py
