+++ src/ray/gcs/gcs_server/test/gcs_actor_scheduler_test.cc
@@ -28,8 +28,7 @@ class GcsActorSchedulerTest : public ::testing::Test {
     gcs_pub_sub_ = std::make_shared<GcsServerMocker::MockGcsPubSub>(redis_client_);
     gcs_table_storage_ = std::make_shared<gcs::RedisGcsTableStorage>(redis_client_);
     gcs_node_manager_ = std::make_shared<gcs::GcsNodeManager>(
-        io_service_, node_info_accessor_, error_info_accessor_, gcs_pub_sub_,
-        gcs_table_storage_);
+        io_service_, error_info_accessor_, gcs_pub_sub_, gcs_table_storage_);
     store_client_ = std::make_shared<gcs::InMemoryStoreClient>(io_service_);
     gcs_actor_table_ =
         std::make_shared<GcsServerMocker::MockedGcsActorTable>(store_client_);
@@ -54,7 +53,6 @@ class GcsActorSchedulerTest : public ::testing::Test {
   std::shared_ptr<gcs::StoreClient> store_client_;
   std::shared_ptr<GcsServerMocker::MockedGcsActorTable> gcs_actor_table_;
 
-  GcsServerMocker::MockedNodeInfoAccessor node_info_accessor_;
   GcsServerMocker::MockedErrorInfoAccessor error_info_accessor_;
 
   std::shared_ptr<GcsServerMocker::MockRayletClient> raylet_client_;
